moodcare-backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==25.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# AI/ML
openai==1.3.0
numpy==1.24.3
pandas==2.0.3

# Cache
django-redis==5.4.0
redis==5.0.1

# WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File handling
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2

# Task Queue
celery==5.3.4

# Utilities
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Push Notifications
firebase-admin==6.4.0